event handler to generate 3 random pokemon working
